requirements.txt
torch == 1.13.0
numpy == 1.23.5
Pillow == 9.3.0
einops == 0.6.0
pandas == 1.5.2
plotly == 5.11.0
matplotlib == 3.6.2
boto3 == 1.26.26
tsp_solver == 0.1
scipy == 1.9.0